backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
python-nmap
shodan
pymetasploit3